multiple webpack configurations is too much overhead recreated single one and moved dist to public folder
