+++ package.json
@@ -22,7 +22,6 @@
         "@types/uglifyjs-webpack-plugin": "^1.1.0",
         "@types/webpack": "^4.4.17",
         "@types/webpack-dev-server": "^3.1.1",
-        "@types/webpack-merge": "^4.1.3",
         "css-loader": "^1.0.1",
         "file-loader": "^1.1.11",
         "prettier": "^1.14.3",
@@ -37,12 +36,11 @@
         "uglifyjs-webpack-plugin": "^1.3.0",
         "webpack": "^4.23.1",
         "webpack-cli": "^3.1.2",
-        "webpack-dev-server": "^3.1.10",
-        "webpack-merge": "^4.1.4"
+        "webpack-dev-server": "^3.1.10"
     },
     "scripts": {
-        "start": "webpack-dev-server --open --config ./webpack.dev.ts",
-        "build": "webpack --config ./webpack.prod.ts"
+        "start": "webpack-dev-server --open --config ./webpack.config.ts --env.production=0",
+        "build": "webpack --config ./webpack.config.ts --env.production=1"
     },
     "author": "Alexander Kaufmann <ak@vivid-planet.com>",
     "license": "ISC"
