+++ public/index.html
@@ -13,7 +13,7 @@
           (function () {
             window.__KWF_REACT_STARTER_ENV__ = 'dev';
             var script = document.createElement('script');
-            script.src = "/assets/build/app.js";
+            script.src = "/assets/build/kwf-react-starter.js";
             script.async = 1;
             document.head.appendChild(script);
           })();
